Moving table data all into one file
