--- playwright-e2e/resources/check_your_answer_content/update_contact_details/consentedUpdateContactDetailsTable.ts
@@ -1,17 +1,0 @@
-export const   consentedUpdateContactDetailsTableData = {
-    tableName: 'Update Contact Details',
-    rows: [
-        { cellItem: "Does this update include a change in representation for either party?", value: "Yes"},
-        "SOLICITOR DETAILS",
-        { cellItem: "Solicitor’s name", value: "Test Baggins" },
-        { cellItem: "Name of your firm", value: "FinRem-1-Org"},
-        { cellItem: "Your reference", value: "Y707HZM"},
-        { cellItem:"Building and Street", value: "3rd Floor, 65-68 Leadenhall St"},
-        { cellItem: "Town or City", value: "London"},
-        { cellItem: "County", value: "Greater London"},
-        { cellItem: "Postcode/Zipcode", value: " EC3A 2AD" },
-        "APPLICANT DETAILS",
-        { cellItem: "Current First and Middle names", value: "Tester case" },
-        { cellItem: "Current Last Name", value: "Baggins" },
-    ]
-}
