+++ playwright-e2e/resources/check_your_answer_content/update_contact_details/updateContactDetailsTable.ts
@@ -21,3 +21,39 @@ export const   contestedUpdateContactDetailsTableData = {
         { cellItem: "Is the Respondent currently a resident in a refuge?", value: "Yes" }
     ]
 }
+
+export const   updateContactDetailsNotRepresentedTable = {
+    tableName: 'Check your answers',
+    rows: [
+        { cellItem: "Does this update include a change in representation for either party?", value: "Yes"},
+        { cellItem: "Select Party to which the change in representation applies:", value: "Respondent" },
+        { cellItem: "Is the respondent represented ?", value: "No" },
+        "RESPONDENT DETAILS",
+        { cellItem: "Current First and Middle names", value: "Smeagol" },
+        { cellItem: "Current Last Name", value: "Gollum" },
+        { cellItem: "Building and Street", value: "10 Selsdon Road" },
+        { cellItem: "Town or City", value: "London" },
+        { cellItem: "Country", value: "United Kingdom" },
+        { cellItem: "Postcode/Zipcode", value: "NW2 7NE" },
+        { cellItem: "Is the Respondent currently a resident in a refuge?", value: "Yes" },
+    ]
+}
+
+export const   consentedUpdateContactDetailsTableData = {
+    tableName: 'Update Contact Details',
+    rows: [
+        { cellItem: "Does this update include a change in representation for either party?", value: "Yes"},
+        "SOLICITOR DETAILS",
+        { cellItem: "Solicitor’s name", value: "Test Baggins" },
+        { cellItem: "Name of your firm", value: "FinRem-1-Org"},
+        { cellItem: "Your reference", value: "Y707HZM"},
+        { cellItem:"Building and Street", value: "3rd Floor, 65-68 Leadenhall St"},
+        { cellItem: "Town or City", value: "London"},
+        { cellItem: "County", value: "Greater London"},
+        { cellItem: "Postcode/Zipcode", value: " EC3A 2AD" },
+        "APPLICANT DETAILS",
+        { cellItem: "Current First and Middle names", value: "Tester case" },
+        { cellItem: "Current Last Name", value: "Baggins" },
+    ]
+}
+
