--- playwright-e2e/resources/check_your_answer_content/update_contact_details/updateContactdetailsNotRepresentedTable.ts
@@ -1,16 +1,0 @@
-export const   updateContactDetailsNotRepresentedTable = {
-    tableName: 'Check your answers',
-    rows: [
-        { cellItem: "Does this update include a change in representation for either party?", value: "Yes"},
-        { cellItem: "Select Party to which the change in representation applies:", value: "Respondent" },
-        { cellItem: "Is the respondent represented ?", value: "No" },
-        "RESPONDENT DETAILS",
-        { cellItem: "Current First and Middle names", value: "Smeagol" },
-        { cellItem: "Current Last Name", value: "Gollum" },
-        { cellItem: "Building and Street", value: "10 Selsdon Road" },
-        { cellItem: "Town or City", value: "London" },
-        { cellItem: "Country", value: "United Kingdom" },
-        { cellItem: "Postcode/Zipcode", value: "NW2 7NE" },
-        { cellItem: "Is the Respondent currently a resident in a refuge?", value: "Yes" },
-    ]
-}
