+++ playwright-e2e/test/consented/update_contact_details.spec.ts
@@ -4,9 +4,9 @@ import { ConsentedEvents } from '../../config/case-data';
 import { YesNoRadioEnum } from '../../pages/helpers/enums/RadioEnums';
 import { updateContactDetailsTabData } from '../../resources/tab_content/consented/update_contact_details_caseworker_tabs';
 import { ConsentedCaseFactory } from '../../data-utils/factory/consented/ConsentedCaseFactory';
-import { consentedUpdateContactDetailsTableData } from '../../resources/check_your_answer_content/update_contact_details/consentedUpdateContactDetailsTable';
+import { consentedUpdateContactDetailsTableData } from '../../resources/check_your_answer_content/update_contact_details/updateContactDetailsTable';
 import { updateRepresentedContactDetailsTabData } from '../../resources/tab_content/consented/update_contact_details_represented';
-import { updateContactDetailsNotRepresentedTable } from '../../resources/check_your_answer_content/update_contact_details/updateContactdetailsNotRepresentedTable';
+import { updateContactDetailsNotRepresentedTable } from '../../resources/check_your_answer_content/update_contact_details/updateContactDetailsTable';
 import { updateNonRepresentedContactDetailsTabData  } from '../../resources/tab_content/consented/update_contact_details_not_represented';
 
 
@@ -54,10 +54,8 @@ test(
       await caseDetailsPage.assertTabData(updateContactDetailsTabData);
     }
 );
-
-
 test(
-    'Consented - Update contact details - change in representation @test',
+    'Consented - Update contact details - change in representation',
     { tag: [] },
     async (
       {
